requirements.txt
numpy==1.24.3
scipy==1.10.1
scikit-learn==1.2.2
transformers>=4.30.0
tqdm>=4.65.0
Pillow>=9.5.0
matplotlib>=3.7.0
pandas>=2.0.0
pyOpenSSL>=0.13,<=24.2.1
pyasn1>=0.1.7
pyasn1-modules>=0.2.1
git+https://github.com/openai/CLIP.git
torch>=2.0.0
torchvision>=0.15.0 